+++ fivox/uriHandler.h
@@ -104,6 +104,21 @@ public:
     float getMagnitude() const;
 
     /**
+     * Get the window of values to consider in the input data for rescaling into
+     * an output data type that is different than float.
+     * The default inputWindow depends on VolumeType:
+     *   (0.0, 2.0) for SPIKES and SYNAPSES
+     *   (-190.0, 0.0) for COMPARTMENTS with TestData, (-80.0, 0.0) otherwise
+     *   (-15.0, 0.0) for SOMAS with TestData, [-80.0, 0.0] otherwise
+     *   (-0.0000147, 0.00225) for LFP with TestData, [-10.0, 10.0] otherwise
+     *   (-100000.0, 300.0) for VSD
+     *
+     * @return a vector containing the minimum and maximum values that the
+     * volume can take before rescaling it
+     */
+    Vector2f getInputWindow() const;
+
+    /**
      * Get the specified path to a dye curve file
      * @return the specified path to the dye curve file
      */